Refined System Logic + Login

+++ src/main/java/Model/Fine.java
@@ -1,5 +1,6 @@
 package Model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Fine {
@@ -8,7 +9,7 @@ public class Fine {
     private Integer loanId;
     private Integer userId;
 
-    private double amount;
+    private BigDecimal amount;
 
     private LocalDate issueDate;
     private LocalDate paymentDate;
@@ -17,7 +18,7 @@ public class Fine {
 
     }
 
-    public Fine(Integer loanId, Integer userId, double amount) {
+    public Fine(Integer loanId, Integer userId, BigDecimal amount) {
         this.loanId = loanId;
         this.userId = userId;
         this.amount = amount;
@@ -28,7 +29,7 @@ public class Fine {
     public Integer getIdFine() { return idFine; }
     public Integer getIdUser() { return userId; }
     public Integer getLoan() { return loanId; }
-    public double getAmount() { return amount; }
+    public BigDecimal getAmount() { return amount; }
     public LocalDate getIssueDate() { return issueDate; }
     public LocalDate getPaymentDate() { return paymentDate; }
 
@@ -36,7 +37,7 @@ public class Fine {
     public void setIdFine(Integer idFine) { this.idFine = idFine; }
     public void setIdLoan(Integer loanId) { this.loanId = loanId; }
     public void setIdUser(Integer userId) { this.userId = userId; }
-    public void setAmount(double amount) { this.amount = amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
     public void setIssueDate(LocalDate issueDate) { this.issueDate = issueDate; }
     public void setPaymentDate(LocalDate paymentDate) { this.paymentDate = paymentDate; }
